requirements.txt
streamlit
pandas
sqlalchemy
pdfplumber==0.10.3
Pillow
streamlit-authenticator==0.2.3
streamlit-sortables
XlsxWriter
supabase
psycopg2-binary
python-dotenv
